+++ index.js
@@ -5,31 +5,35 @@ require('dotenv').config();
 
 const cors = require('cors');
 
+const cal = require('./api/googleCal');
+
 const app = express();
 app.use(express.json());
-
 app.use(express.urlencoded({ extended: false }));
 app.use(morgan('dev'));
 
-app.use(cors())
-
 app.get('/', async (req, res, next) => {
-  res.send({ message: 'Awesome it works 🐻' });
+    res.send({ message: 'Awesome it works 🐻' });
 });
 
-app.use('/api', require('./routes/api.route'));
+app.use(cors())
+
+app.use("/api/cal", cal);
 
 app.use((req, res, next) => {
-  next(createError.NotFound());
+    next(createError.NotFound());
 });
 
 app.use((err, req, res, next) => {
-  res.status(err.status || 500);
-  res.send({
-    status: err.status || 500,
-    message: err.message,
-  });
+    res.status(err.status || 500);
+    res.send({
+      status: err.status || 500,
+      message: err.message,
+    });
 });
 
 const PORT = process.env.PORT || 3001;
-app.listen(PORT, () => console.log(`🚀 Starting Server on Port ${PORT}`));
+
+app.listen(PORT, () => {
+    console.log(`Starting Server on Port ${PORT}`);
+})
