+++ package.json
@@ -2,12 +2,12 @@
   "name": "jnce-calendar-api",
   "version": "1.0.0",
   "description": "",
-  "main": "app.js",
+  "main": "index.js",
   "engines": {
     "node": "14.x"
   },
   "scripts": {
-    "start": "node app.js",
+    "start": "node index.js",
     "dev": "nodemon app.js"
   },
   "keywords": [],
