+++ vercel.json
@@ -2,7 +2,7 @@
     "version": 2,
     "builds": [
         {
-            "src": "./app.js",
+            "src": "./index.js",
             "use": "@vercel/node"
         }
     ],
